requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
dnspython==2.4.2
python-whois==0.8.0
pydantic-settings==2.1.0
pandas==2.1.3
openpyxl==3.1.2
python-multipart==0.0.6
idna==3.6
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0
flake8==6.1.0
black==23.12.1
mypy==1.8.0
# Celery and Redis for async queue
celery==5.3.4
redis==5.0.1
# PDF generation (G17)
reportlab==4.0.7
# Security (G19 pre-flight)
bcrypt>=4.0.0
structlog>=23.0.0
sentry-sdk[fastapi]>=1.38.0
# Microsoft SSO (G19)
msal>=1.24.0
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0  # For refresh token encryption (Fernet)
# Retry logic for deadlock prevention (P1-4)
tenacity>=8.2.3

